requirements.txt
Django==4.2.23
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-extensions==4.1
python-decouple==3.8
Pillow==10.4.0
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
psycopg2-binary==2.9.10
django-filter==24.1
